+++ ts/src/components.ts
@@ -1,4 +1,0 @@
-export * from './components/dialog.component';
-export * from './components/dropdown.component';
-export * from './components/date-picker.component';
-export * from './components/grid';