+++ ts/src/components/index.ts
@@ -2,3 +2,4 @@ export * from './dialog.component';
 export * from './dropdown.component';
 export * from './date-picker.component';
 export * from './grid';
+export * from './form';