+++ ts/src/core/index.ts
@@ -1,3 +1,7 @@
 export * from './types';
 export * from './cache';
 export * from './ajax';
+export * from './observable';
+
+import * as _Rosie from './utils';
+export const Rosie = { ..._Rosie };