+++ ts/src/index.ts
@@ -2,8 +2,3 @@ import './lang/string';
 import './lang/number';
 import './lang/date';
 import './lang/array';
-
-export * from './core';
-
-import * as _Rosie from './util';
-export const Rosie = { ..._Rosie };